Add shaker library
Description
This commit add the library with the RLPx handshake functionality.

Implementation details
The ECIES operation are contained in the ecies modules, this
module is used to encrypt and decrypt data communicated between the
initiator and the recipient, the messages supported can be found in the
message module. The message authenticator is contained in
the mac module, this module help the ecies modules with encrypting
messages. The secrets used by the ecies operatios are contained secrets module. Library errors can be found in the errors module. In the shaker module the higher level Shaker struct can be found which contains the functionality to decrypt the ecies secrets and the reading and writing of frames. Finally there is the stream module which will be used when writing clients to initialize a connection with an Ethereum node.